refactor(core): keep arrow uniform mappings in WebGL
Remove GLSL ordering details from the semantic Arrow mark.

Keep only the head-shape and placement mappings consumed by the WebGL delegate.
